feat(core/cli): add CLI platform!
In this commit, I have introduced a new part of the framework for the CLI builder. It is similar in
concept to the Server builder, but instead of importing middlewares and controllers, you will import
Commands.

Now bootstrap functions are contained in their platform functions. To upgrade, just change
`bootstrap(AppModule)` to `platformServer().bootstrap(AppModule)`, and import `platformServer` from
`@orbital/core`.